backend/requirements.txt
# Core dependencies
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
python-dotenv==1.2.1
python-multipart==0.0.21

# Azure OpenAI SDK
openai>=1.0.0

# Authentication
bcrypt==4.1.3
PyJWT==2.10.1
python-jose==3.5.0
passlib==1.7.4

# Data validation
pydantic==2.12.5
pydantic_core==2.41.5
email-validator==2.3.0

# PDF processing
PyPDF2==3.0.1

# Async support
anyio==4.12.0
aiohttp==3.13.2

# HTTP client
httpx==0.28.1
requests==2.32.5

# Utilities
typing_extensions==4.15.0
python-dateutil==2.9.0.post0
